requirements-data.txt
pyspark==3.5.0
minio==7.2.0
requests==2.31.0
beautifulsoup4==4.12.2
tqdm==4.66.1
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.26.2
aiohttp==3.9.1
asyncio==3.4.3
apache-airflow==2.7.3
apache-airflow-providers-apache-spark==4.1.3
apache-airflow-providers-http==4.5.1
apache-airflow-providers-ssh==3.7.1
great-expectations==0.18.11
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
streamlit==1.29.0
plotly==5.18.0
sqlalchemy-utils==0.41.1
alembic==1.13.1 